Rehaul of the RMarkdown files
I figured that having a bunch of nested lists was inviting mayhem, so everything is now running from two .csv files that include the aperture and the species data, respectively.

I also updated some of the code to be shorter and more reproducible using tidyverse grammar.

Lastly, I also updated all the file paths using the here() package, which should help to run the data on other machines.